using transforms from examples directory for code re-use
